all compile-time errors eliminated, going to bug-detection phase..

+++ src/core/camera.h
@@ -35,8 +35,8 @@ public:
 
 	KRR_CALLABLE Ray getRay(vec2i pixel, vec2i frameSize, Sampler sampler) {
 		Ray ray;
-		vec2f p = vec2f(pixel.cast<float>()) + vec2f(0.5f) + sampler.get2D() /*uniform sample + box filter*/;
-		vec2f ndc = vec2f(2 * p) / vec2f(frameSize.cast<float>()) + vec2f(-1.f); // ndc in [-1, 1]^2
+		vec2f p = (vec2f)pixel + vec2f(0.5f) + sampler.get2D() /*uniform sample + box filter*/;
+		vec2f ndc = vec2f(2 * p) / vec2f(frameSize) + vec2f(-1.f); // ndc in [-1, 1]^2
 		if (mData.lensRadius > 0) {			/*Thin lens*/
 			vec3f focalPoint = mData.pos + ndc[0] * mData.u + ndc[1] * mData.v + mData.w;
 			vec2f apertureSample = mData.lensRadius > M_EPSILON ? uniformSampleDisk(sampler.get2D()) : vec2f::Zero();
